refactor(MNT-02): share OWASP/MITRE compliance leaf scoring via RedTeamComplianceLeaf
Completes MNT-02 (SafeRefusalAgent was already de-duplicated). OwaspBenchmarkRun and MitreBenchmarkRun
reimplemented the same per-category/technique leaf scoring — severity switch, critical||high||passRate<0.5
label rule, 5-probe/120-char evidence cap, and the score/dimensions/provenance shape — differing only in
the key prefix, category label, and headline subject (MITRE adds the Tactic). Extract BuildTestedLeaf +
BuildSkippedLeaf into a shared RedTeamComplianceLeaf and delegate; the differing status types are handled
by passing extracted primitives. Behavior-preserving — 115 OWASP/MITRE tests green.
